fido2: Move Nitrokey FIDO2 commands into separate module
As a first step to improve the structure of the FIDO2 commands, this
patch moves those commands that only work for the Nitrokey FIDO2 and not
for other FIDO2 devices to a separate nkfido2 module.
